front end components added
