+++ src/components/Home.jsx
@@ -1,0 +1,19 @@
+import React, { useState, useEffect } from 'react';
+
+
+const Home = () => {
+ 
+  
+    useEffect(() => {
+ 
+    }, []);
+  
+    return (
+      <div className='app-container'>
+        <h1>Home Page!</h1>
+        <p>API Status:</p>
+      </div>
+    );
+  };
+  
+  export default Home;
