+++ src/components/index.jsx
@@ -1,0 +1,6 @@
+export { default as App } from './App';
+export { default as Cars } from './Cars';
+export { default as Cart } from './Cart';
+export { default as Home } from './Home';
+export { default as Login } from './Login';
+export { default as Register } from './Register';
